requirements-dev.txt
autoflake
coverage
black
bump2version
codespell
flake8
isort
mypy
pip
pytest
pytest-runner
pyyaml
sphinx
sphinx-autodoc-typehints
sphinx-rtd-theme
tox
twine
wheel
